% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{constructLD}
\alias{constructLD}
\title{Construct LD matrix from data.table}
\usage{
constructLD(dfld, ids)
}
\arguments{
\item{dfld}{\code{data.table} storing LD information}

\item{ids}{variant names to include}
}
\value{
\code{sparseMatrix} storing LD between variants
}
\description{
Construct LD matrix from data.table
}
