% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lambda.R
\name{estimate_lambda}
\alias{estimate_lambda}
\title{Estimate lambda shrinkage parameter}
\usage{
estimate_lambda(
  X,
  method = c("Ledoit-Wolf", "OAS", "Touloumis", "Schafer-Strimmer")
)
}
\arguments{
\item{X}{matrix with features as columns}

\item{method}{method used to estimate lambda shrinkage parameter}
}
\value{
estimated lambda value
}
\description{
Estimate lambda shrinkage parameter with specified method
}
\examples{
n = 100
p = 400

X = matrix(rnorm(n*p), n, p)

estimate_lambda(X, "Ledoit-Wolf")

}
