% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLDMatrix.R
\name{fillRate}
\alias{fillRate}
\title{Compute fill rate of sparse matrix}
\usage{
fillRate(M)
}
\arguments{
\item{M}{\code{sparseMatrix}}
}
\value{
fraction of entries that are non-zero
}
\description{
Compute fill rate of sparse matrix
}
