% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_region.R
\name{impute_region}
\alias{impute_region}
\title{Impute many z-statistics}
\usage{
impute_region(
  df,
  gds,
  region,
  flankWidth,
  method = c("decorrelate", "Ledoit-Wolf", "OAS", "Touloumis", "Schafer-Strimmer"),
  lambda = NULL
)
}
\arguments{
\item{df}{\code{data.frame} with columns \code{ID}, \code{z}, \code{GWAS_A1}, \code{GWAS_A2}, \code{chrom}, \code{position} \code{A1}, \code{A2}.}

\item{gds}{\code{GenomicDataStream} of reference panel}

\item{region}{genomic region to impute}

\item{flankWidth}{additional window added to \code{region}}

\item{method}{method used to estimate shrinkage parameter lambda.  default is \code{"decorrelate"}}

\item{lambda}{(default: NULL) value used to shrink correlation matrix}
}
\description{
Impute many z-statistics given observed z-statistics and reference panel
}
