% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{impute_z}
\alias{impute_z}
\title{Impute z-score based correlations}
\usage{
impute_z(z, Sigma, i, lambda = 0.1)
}
\arguments{
\item{z}{vector of observed z-scores}

\item{Sigma}{matrix of correlation between z-scores}

\item{i}{index of z-scores to impute}

\item{lambda}{value used to shrink correlation matrix}
}
\value{
\code{data.frame} with variant \code{ID}, imputed \code{z.stat} for test \code{i}, standard error \code{se} of the impute z-score, and metric of accuracy \code{r2.pred}.
}
\description{
Impute z-score from a missing test based on z-scores for other tests, and the correlation matrix between z-scores
}
