% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.eclairs.R
\name{imputez.eclairs}
\alias{imputez.eclairs}
\title{Impute z-score based correlations}
\usage{
imputez.eclairs(z, X, i, lambda = NULL)
}
\arguments{
\item{z}{vector of observed z-scores}

\item{X}{matrix genotype data from a reference panel}

\item{i}{index of z-scores to impute}

\item{lambda}{(default: NULL) value used to shrink correlation matrix}
}
\value{
\code{data.frame} storing:
\describe{
  \item{ID}{variant identifier}
  \item{z.stat}{imputed z-statistic}
  \item{sigSq}{variance of imputed z-statistic}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{lambda}{correlation shrinkage parameter}
}
}
\description{
Impute z-score from a missing test based on z-scores for other tests, and the genotype matrix
}
\details{
Uses implicit covariance and emprical Bayes shrinkage of the eigen-values to accelerate computations. \code{imputez()} is cubic in the number of features, p, but \code{imputez.eclairs()}  is the minimum of O(n p^2) and O(n^2p).  For large number of features, this is can be a dramatic speedup.
}
\references{
{
  \insertRef{pasaniuc2014fast}{imputez}
}
}
\seealso{
\code{imputez}
}
