% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputezDecorr.R
\name{imputezDecorr}
\alias{imputezDecorr}
\title{Impute z-score based correlations}
\usage{
imputezDecorr(z, X, i, k = min(nrow(X), ncol(X) - length(i)), lambda = NULL)
}
\arguments{
\item{z}{vector of observed z-scores}

\item{X}{matrix genotype data from a reference panel}

\item{i}{index of z-scores to impute}

\item{k}{rank of SVD}

\item{lambda}{(default: NULL) value used to shrink correlation matrix}
}
\value{
\code{data.frame} storing:
\describe{
  \item{ID}{variant identifier}
  \item{z.stat}{imputed z-statistic}
  \item{se}{standard error of imputed z-statistic}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{lambda}{shrinkage parameter}
}
}
\description{
Impute z-score from a missing test based on z-scores for other tests, and the genotype matrix
}
\details{
Uses implicit covariance and emprical Bayes shrinkage of the eigen-values to accelerate computations. \code{imputez()} is cubic in the number of features, p, but \code{imputezDecorr()}  is the minimum of O(n p^2) and O(n^2p).  For large number of features, this is can be a dramatic speedup.
}
\references{
Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.
}
\seealso{
\code{imputez()}, \code{decorrelate::eclairs()}, \code{decorrelate::decorrelate()}
}
