% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{pairwiseCompleteWindow}
\alias{pairwiseCompleteWindow}
\title{Get complete subset of correlation matrix}
\usage{
pairwiseCompleteWindow(S, mid)
}
\arguments{
\item{S}{correlation matrix}

\item{mid}{index of target variant}
}
\value{
begin and end indeces of completely filled matrixq
}
\description{
Get complete subset of correlation matrix so that no entries are zero
}
