% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLDMatrix.R
\name{readLDMatrix}
\alias{readLDMatrix}
\title{Read plink LD files}
\usage{
readLDMatrix(df_files)
}
\arguments{
\item{df_files}{data.frame of file info: chrom, BIM, LD}
}
\value{
list with one entry per chromosome with \code{C.ld} storing the LD correlation matrix and \code{gr_position} storing the variant information
}
\description{
Read LD files generated by plink into a list of sparseMatrix for each chromosome
}
