% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Impute many z-statistics}
\usage{
run_analysis(
  targets,
  df,
  gds,
  window = 1e+05,
  lambda = NULL,
  useEclairs = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{targets}{variant ID's to impute}

\item{df}{\code{data.frame} with columns \code{ID}, \code{z}, \code{GWAS_A1}, \code{GWAS_A2}, \code{chrom}, \code{position} \code{A1}, \code{A2}.}

\item{gds}{\code{GenomicDataStream} of reference panel}

\item{window}{size of window in bp}

\item{lambda}{shrinkage parameter used for LD matrix.  Defaults to \code{lambda = NULL} and is estimated from the data}

\item{useEclairs}{default \code{TRUE} to use faster method.}

\item{quiet}{suppress messages}
}
\value{
\code{data.frame} storing imputed results:
\describe{
  \item{ID}{variant identifier}
  \item{z.stat}{imputed z-statistic}
  \item{sigSq}{variance of imputed z-statistic}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{lambda}{shrinkage parameter}
  \item{maf}{minor allele frequency in reference panel}
  \item{nVariants}{number of variants used in imputation}
}
}
\description{
Impute many z-statistics given observed z-statistics and reference panel
}
\details{
Implements method by Pasaniuc, et al. (2014).
}
\references{
Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.
}
\seealso{
\code{imputez()}
}
