% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputability.R
\name{run_imputability}
\alias{run_imputability}
\title{Evaluate imputation quality}
\usage{
run_imputability(
  df,
  gds,
  window,
  flankWidth,
  ids,
  method = c("decorrelate", "Ledoit-Wolf", "OAS", "Touloumis", "Schafer-Strimmer"),
  lambda = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{\code{data.frame} with columns \code{ID},  \code{CHROM}, \code{POS}}

\item{gds}{\code{GenomicDataStream} of reference panel}

\item{window}{size of window in bp}

\item{flankWidth}{additional window added to \code{region}}

\item{ids}{variant IDs to evaluate imputation r2 score}

\item{method}{method used to estimate shrinkage parameter lambda.  default is \code{"decorrelate"}}

\item{lambda}{(default: NULL) value used to shrink correlation matrix. Only used if method is \code{"decorrelate"}}

\item{quiet}{suppress messages}

\item{...}{additional arguments passed to \code{impute_region()}}
}
\value{
\code{tibble} storing imputed results:
\describe{
  \item{ID}{variant identifier}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{lambda}{shrinkage parameter}
  \item{maf}{minor allele frequency in reference panel}
  \item{nVariants}{number of variants used in imputation}
}
}
\description{
Evaluate
}
\details{
Implements method by Pasaniuc, et al. (2014).
}
\examples{
library(GenomicDataStream)
library(tidyverse)

# VCF file for reference
file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")

# initialize data stream
gds <- GenomicDataStream(file, "DS", initialize=TRUE)

# read file of variant locations
file <- system.file("extdata", "test.map", package = "GenomicDataStream")
df = read_tsv(file, show_col_types=FALSE)

# evaluate imputation r2 for these variants
ids = df$ID[c(1,9)]

# Given GenomicDataStream of reference panel,
# compute imputation r2 for variants in ids
run_imputability(df, gds, 10000, 1000, ids)
}
\references{
Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.
}
\seealso{
\code{run_imputez()}
}
