% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{imputez}
\alias{imputez}
\title{Impute z-score based correlations}
\usage{
imputez(z, Sigma, i, lambda = 0.1, useginv = FALSE)
}
\arguments{
\item{z}{vector of observed z-scores}

\item{Sigma}{matrix of correlation between z-scores}

\item{i}{index of z-scores to impute}

\item{lambda}{value used to shrink correlation matrix}

\item{useginv}{if \code{TRUE} use pseudoinverse}
}
\value{
\code{data.frame} storing:
\describe{
  \item{ID}{variant identifier}
  \item{z.stat}{imputed z-statistic}
  \item{se}{standard error of imputed z-statistic}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{lambda}{shrinkage parameter}
}
}
\description{
Impute z-score from a missing test based on z-scores for other tests, and the correlation matrix between z-scores
}
\details{
Implements method by Pasaniuc, et al. (2014).
}
\examples{
library(GenomicDataStream)
library(mvtnorm)

# VCF file for reference
file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")

# initialize data stream
gds <- GenomicDataStream(file, "DS", initialize=TRUE)

# read genotype data from reference
dat <- getNextChunk(gds)

# simulate z-statistics with correlation structure
# from the LD of the reference panel
C <- cor(dat$X)
set.seed(1)
z <- c(rmvnorm(1, rep(0, 10), C))
names(z) <- colnames(dat$X)

# Impute z-statistics for variants 2 and 3 
# using the other variants and observed z-statistics
# from the reference panel
imputez(z, C, 2:3)
}
\references{
Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.
}
