% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{run_imputez}
\alias{run_imputez}
\title{Impute many z-statistics}
\usage{
run_imputez(z, LDinfo, IDs, maxWindowSize = 200, quiet = FALSE)
}
\arguments{
\item{z}{vector of observed z-statistics with unobserved storing value \code{NA}}

\item{LDinfo}{\code{list} with \code{data.table} storing LD information, and \code{GRanges} with position info}

\item{IDs}{variant names in \code{z} to impute}

\item{maxWindowSize}{max window size around target variant to keep}

\item{quiet}{default FALSE.  If TRUE, suppress progress bar}
}
\value{
\code{data.frame} storing imputed results:
\describe{
  \item{ID}{variant identifier}
  \item{z.stat}{imputed z-statistic}
  \item{sigSq}{variance of imputed z-statistic}
  \item{r2.pred}{metric of accuracy of the imputed z-statistic based on its variance}
  \item{width}{number of SNPs in LD panel used in imputation}
}
}
\description{
Impute many z-statistics given observed z-statistics and \code{data.table} of LD.
}
\details{
Implements method by Pasaniuc, et al. (2014).
}
\references{
{
  \insertRef{pasaniuc2014fast}{imputez}
}
}
\seealso{
\code{imputez()}
}
