% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputez.R
\name{run_imputez}
\alias{run_imputez}
\title{Impute many z-statistics}
\usage{
run_imputez(z, dfld, idx, maxWindowSize = 200)
}
\arguments{
\item{z}{vector of observed z-statistics with unobserved storing value \code{NA}}

\item{dfld}{\code{data.table} storing LD information}

\item{idx}{indeces of \code{z} to impute}

\item{maxWindowSize}{max window size around target variant to keep}
}
\value{
\code{data.frame} storing imputed results
}
\description{
Impute many z-statistics given observed z-statistics and \code{data.table} of LD.
}
\details{
Implements method by Pasaniuc, et al. (2014).
}
\references{
{
  \insertRef{pasaniuc2014fast}{imputez}
}
}
